[refactor] Merge lex into parse
