+++ zydeco-lang/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod lex;
 pub mod parse;
 pub mod statics;
 pub mod dynamics;
