+++ zydeco-lang/src/parse.rs
@@ -3,14 +3,29 @@ lalrpop_mod!(pub parser, "/parse/parser.rs");
 
 pub mod fmt;
 pub mod syntax;
+pub mod token;
 
 pub use parser::{ExpressionParser, ZydecoParser};
 
-use std::path::PathBuf;
+use logos::{Logos, SpannedIter};
+use token::Tok;
 
-pub struct ZydecoFile {
-    pub path: PathBuf,
-    pub content: String,
+pub struct Lexer<'source> {
+    inner: SpannedIter<'source, Tok<'source>>,
+}
+
+impl<'source> Lexer<'source> {
+    pub fn new(source: &'source str) -> Self {
+        Self { inner: Tok::lexer(&source).spanned() }
+    }
+}
+
+impl<'source> Iterator for Lexer<'source> {
+    type Item = (usize, Tok<'source>, usize);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.inner.next().map(|(tok, range)| (range.start, tok, range.end))
+    }
 }
 
 /// Expand escape characters in a string literal, converting the source code
@@ -34,7 +49,7 @@ pub fn apply_string_escapes(code: &str) -> String {
                     'n' => '\n',
                     'r' => '\r',
                     't' => '\t',
-                    _ => next_ch
+                    _ => next_ch,
                 }
             }
             text.push(ch);
@@ -43,8 +58,8 @@ pub fn apply_string_escapes(code: &str) -> String {
     }
 }
 
-
-pub fn apply_char_escapes(s: &str) -> char { //lib function:Logos
+pub fn apply_char_escapes(s: &str) -> char {
+    //lib function:Logos
     let ch = s.chars().nth(1).unwrap();
     match ch {
         'n' => '\n',
