+++ zydeco-lang/src/parse/parser.lalrpop
@@ -1,11 +1,9 @@
 use crate::{
-    lex::token::*,
-    parse::syntax::*,
-    syntax::{binders::*, TCtor},
-    syntax::ann::*,
+    parse::{syntax::*, token::*},
+    syntax::{ann::*, binders::*, TCtor},
 };
-use crate::parse::apply_string_escapes;
-use crate::parse::apply_char_escapes;
+use crate::parse::{apply_string_escapes, apply_char_escapes};
+
 grammar<'input>(input: &'input str);
